ASoC: wcd: move cap setting to codec driver
Cap setting is needed for any micbias and so it should be done
from codec driver instead of mbhc driver which uses only micbias2.
Codec driver will read the device tree to find out if external cap
is present or not and will set the bit for cap setting accordingly.

Change-Id: I20d2a104915977929a591260bcd0ea9a48ad475b
Signed-off-by: Simmi Pateriya <simmip@codeaurora.org>
Signed-off-by: Yeleswarapu Nagaradhesh <nagaradh@codeaurora.org>

Conflicts:
	sound/soc/codecs/msm8x16-wcd.h